chat-django/requirements.txt
Django>=5.0,<6.0
django-cors-headers>=4.3
python-dotenv>=1.2.1
httpx>=0.28.1
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
numpy>=1.26.4